backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
telethon
PySocks
google-generativeai
openai
pdfplumber
pymupdf
pytesseract
pdf2image
Pillow
python-docx
beautifulsoup4
httpx
python-multipart
